Show all books under item section (in progress)
I uncommented the line under "show all items" and removed the LIMIT 1 parameter. It's now showing 3 out of 10 books under Items instead of just 1. I'm not sure why the other 7 aren't showing up so I'll be continuing to work on this, but I wanted to get this progress uploaded for now.